Update for latest python-ihm/python-modelcif
Add the original PDB chain ID of the template,
as the 'strand_id' argument to the Template
constructor, to populate the
_ma_template_details.template_auth_asym_id
field. Update known-good mmCIF files to include
this mandatory field, so that they validate
against the dictionary again. Also, extend
each test model and template to include at
least 2 residues so that python-ihm does not
treat them as non-polymers.